move unit test to test_dtensor_compile.py
